feat(analytics): Rename component.
