+++ components/analytics/tracking-events.ts
@@ -18,7 +18,7 @@ interface duplexClickedProps extends duplexViewedProps {}
 export type EventName = keyof EventsMap;
 export type EventData<T extends keyof EventsMap> = EventsMap[T];
 
-export function createAnalyticEvent<T extends EventName>(eventName: T, eventData: EventData<T>) {
+export function createAnalyticsEvent<T extends EventName>(eventName: T, eventData: EventData<T>) {
   return {
     eventName,
     eventData,
