+++ components/duplex-ctf/duplex-ctf-client.tsx
@@ -3,30 +3,33 @@
 import { ResultOf } from 'gql.tada';
 import { useAnalytics } from 'use-analytics';
 
+import { createAnalyticsEvent } from '#/components/analytics/tracking-events';
+import { TrackInView } from '#/components/analytics/trackInView';
 import { RichTextCtf } from '#/components/rich-text-ctf';
 
-import { createAnalyticEvent } from '../analytic/tracking-events';
-import { TrackInView } from '../analytic/trackInView';
 import { useComponentPreview } from '../hooks/use-component-preview';
 import { getImageChildProps } from '../image-ctf';
 import { getPageLinkChildProps } from '../page';
 import { Duplex } from '../ui/duplex';
 import { ComponentDuplexFieldsFragment } from './duplex-ctf';
 
-// We can create analytic event typed data on top level
-// using createAnalyticEvent helper.
+// We can create analytics event typed data on top level
+// using createAnalyticsEvent helper.
 // It will us type-safely create event name and event data.
-const analyticInViewEvent = createAnalyticEvent('duplexViewed', {
+const analyticsInViewEvent = createAnalyticsEvent('duplexViewed', {
   category: 'duplexViewed',
   type: 'ctf',
 });
 
 // For the direct track() usage from hook we can destructure the object
-// returned from createAnalyticEvent helper to have separate values as props.
-const { eventName: analyticClickEventName, eventData: analyticClickEventData } = createAnalyticEvent('duplexClicked', {
-  category: 'duplexClicked',
-  type: 'ctf',
-});
+// returned from createAnalyticsEvent helper to have separate values as props.
+const { eventName: analyticsClickEventName, eventData: analyticsClickEventData } = createAnalyticsEvent(
+  'duplexClicked',
+  {
+    category: 'duplexClicked',
+    type: 'ctf',
+  }
+);
 
 export const DuplexCtfClient: React.FC<{
   data: ResultOf<typeof ComponentDuplexFieldsFragment>;
@@ -35,7 +38,7 @@ export const DuplexCtfClient: React.FC<{
   const { data, addAttributes } = useComponentPreview<typeof originalData>(originalData);
   const { track } = useAnalytics();
   return (
-    <TrackInView {...analyticInViewEvent}>
+    <TrackInView {...analyticsInViewEvent}>
       <Duplex
         headline={data.headline}
         bodyText={
@@ -64,7 +67,7 @@ export const DuplexCtfClient: React.FC<{
           })
         }
         colorPalette={data.colorPalette}
-        onClickAnalyticEvent={() => track(analyticClickEventName, analyticClickEventData)}
+        onClickAnalyticsEvent={() => track(analyticsClickEventName, analyticsClickEventData)}
       />
     </TrackInView>
   );
