+++ components/hero-banner-ctf/hero-banner-ctf-client.tsx
@@ -2,11 +2,11 @@
 
 import { ResultOf } from 'gql.tada';
 
+import { createAnalyticsEvent } from '#/components/analytics/tracking-events';
+import { TrackInView } from '#/components/analytics/trackInView';
 import { ComponentHeroBannerFieldsFragment } from '#/components/hero-banner-ctf/hero-banner-ctf';
 import { RichTextCtf } from '#/components/rich-text-ctf';
 
-import { createAnalyticEvent } from '../analytic/tracking-events';
-import { TrackInView } from '../analytic/trackInView';
 import { useComponentPreview } from '../hooks/use-component-preview';
 import { getImageChildProps } from '../image-ctf';
 import { getPageLinkChildProps } from '../page';
@@ -17,12 +17,12 @@ export const HeroBannerCtfClient: React.FC<{
 }> = (props) => {
   const { data: originalData } = props;
   const { data, addAttributes } = useComponentPreview<typeof originalData>(originalData);
-  // We use createAnalyticEvent helper to create typed event.
-  const analyticInViewEvent = createAnalyticEvent('heroBannerViewed', {
+  // We use createAnalyticsEvent helper to create typed event.
+  const analyticsInViewEvent = createAnalyticsEvent('heroBannerViewed', {
     category: 'duplexViewed',
   });
   return (
-    <TrackInView {...analyticInViewEvent}>
+    <TrackInView {...analyticsInViewEvent}>
       <HeroBanner
         headline={data.headline}
         bodyText={data.bodyText && <RichTextCtf {...data.bodyText} />}
